bummer: for at least some flowcells, need to go to u32, ~doubling memory usage.

+++ src/main.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use log::{info, warn, error};
 
 struct PrimaryKeyMap {
-    map: HashMap<String, u16>,
+    map: HashMap<String, u32>,
     reverse_map: Vec<String>,
 }
 
@@ -15,12 +15,12 @@ impl PrimaryKeyMap {
         }
     }
 
-    fn key_for_value(&mut self, key: &str) -> u16 {
+    fn key_for_value(&mut self, key: &str) -> u32 {
         match self.map.get(key) {
             Some(value) => *value,
             None => {
-                let new_value = self.reverse_map.len() as u16;
-                if new_value == std::u16::MAX {
+                let new_value = self.reverse_map.len() as u32;
+                if new_value == std::u32::MAX {
                     panic!("Too many distinct values in PrimaryKeyMap");
                 }
                 self.map.insert(key.to_string(), new_value);
@@ -30,14 +30,14 @@ impl PrimaryKeyMap {
         }
     }
 
-    fn value_for_key(&self, key: u16) -> Option<&String> {
+    fn value_for_key(&self, key: u32) -> Option<&String> {
         self.reverse_map.get(key as usize)
     }
 }
 
-fn name_to_readname_key(name: &str, primary_key_map: &mut PrimaryKeyMap) -> [u16; 4] {
+fn name_to_readname_key(name: &str, primary_key_map: &mut PrimaryKeyMap) -> [u32; 4] {
     let name_parts = name.split(':').collect::<Vec<&str>>();
-    let mut key = [0 as u16; 4];
+    let mut key = [0 as u32; 4];
     key[0] = primary_key_map.key_for_value(&name_parts[0..4].join(":"));
     for i in 0..3 {
         key[i + 1] = primary_key_map.key_for_value(name_parts[4+i]);
@@ -49,7 +49,7 @@ fn main() {
     env_logger::init();
     let mut readname_part_pkm = PrimaryKeyMap::new();
     let mut comment_pkm = PrimaryKeyMap::new();
-    let mut readnames_to_comments: HashMap<[u16; 4], u16> = HashMap::new();
+    let mut readnames_to_comments: HashMap<[u32; 4], u32> = HashMap::new();
     let mut args = std::env::args().collect::<Vec<String>>();
 
     if args.len() < 2 {
